loading minisearch from serialized json
