+++ gatsby/src/components/search-engine/useSearchEngine.tsx
@@ -1,7 +1,9 @@
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
 import { ISearchIndexQuery } from '@graphql-types';
-import { SearchResult } from 'minisearch';
+import removeAccents from 'remove-accents';
+import MiniSearch, { SearchResult } from 'minisearch';
+import sanitizeHtml from 'sanitize-html';
 
 export const query = graphql`
   query SearchIndex {
@@ -25,14 +27,37 @@ export type Result = IndexedResult & SearchResult;
 const useSearchEngine = () => {
   const [results, setResults] = useState<Result[]>([]);
   const data = useStaticQuery<ISearchIndexQuery>(query);
+  const miniSearch = useMemo(
+    () =>
+      MiniSearch.loadJSON<Result>(
+        JSON.stringify(data.siteMiniSearchIndex.index),
+        {
+          fields: ['title', 'content'],
+          processTerm: (term) => {
+            const withoutAccents = removeAccents.remove(term);
+            const lowerCase = withoutAccents.toLocaleLowerCase();
+            const withoutHtml = sanitizeHtml(lowerCase, { allowedTags: [] });
+            return withoutHtml;
+          },
+          searchOptions: {
+            processTerm: (term) => {
+              return removeAccents.remove(term).toLocaleLowerCase();
+            },
+            boost: { title: 2, content: 1 },
+            prefix: true,
+            fuzzy: 0.2,
+          },
+        },
+      ),
+    [],
+  );
 
   const handleSearch = (term: string) => {
     if (data) {
-      const results = data.siteMiniSearchIndex.index.search(term) as Result[];
+      const results = miniSearch.search(term) as Result[];
       setResults(results);
     }
   };
-
   return [handleSearch, results] as const;
 };
